printforge/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "printforge"
version = "0.1.0"
description = "3D Printer Control System"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pyserial>=3.5",
    "pyserial-asyncio>=0.6",
    "aiosqlite>=0.19",
    "httpx>=0.27",
    "python-multipart>=0.0.9",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
